+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * Changed connection secure to true by default
 * Renamed public package `balancer` tp `balancers` (this package contains only constructors of balancers)
 * Moved interfaces from package `internal/balancer/ibalancer` to `internal/balancer`
+* Added `NextResultSetErr()` func for select next result set and return error
+* Added package `table/result/indexed` with interface `indexed.Value`
+* Replaced abstract `interface{}` in `Scan` and `ScanWithDefaults` to `indexed.Value`
+* Splitted `trace.Table.OnPoolRetry` callback to `trace.Table.OnPoolDo` and `trace.Table.OnPoolDoTx` callbacks
 
 ## 3.7.2
 * Retry remove directory in `sugar.RemoveRecursive()` for retryable error